Crate rusoto_eks

Amazon Elastic Container Service for Kubernetes (Amazon EKS) is a managed service that makes it easy for you to run Kubernetes on AWS without needing to stand up or maintain your own Kubernetes control plane. Kubernetes is an open-source system for automating the deployment, scaling, and management of containerized applications.

Amazon EKS runs up-to-date versions of the open-source Kubernetes software, so you can use all the existing plugins and tooling from the Kubernetes community. Applications running on Amazon EKS are fully compatible with applications running on any standard Kubernetes environment, whether running in on-premises data centers or public clouds. This means that you can easily migrate any standard Kubernetes application to Amazon EKS without any code modification required.

If you're using the service, you're probably looking for EksClient and Eks.

Structs

Certificate
An object representing the certificate-authority-data for your cluster.

Cluster
An object representing an Amazon EKS cluster.

CreateClusterRequest

CreateClusterResponse

DeleteClusterRequest

DeleteClusterResponse

DescribeClusterRequest

DescribeClusterResponse

DescribeUpdateRequest

DescribeUpdateResponse

EksClient
A client for the Amazon EKS API.

ErrorDetail
An object representing an error when an asynchronous operation fails.

ListClustersRequest

ListClustersResponse

ListUpdatesRequest

ListUpdatesResponse

LogSetup
An object representing the enabled or disabled Kubernetes control plane logs for your cluster.

Logging
An object representing the logging configuration for resources in your cluster.

Update
An object representing an asynchronous update.

UpdateClusterConfigRequest

UpdateClusterConfigResponse

UpdateClusterVersionRequest

UpdateClusterVersionResponse

UpdateParam
An object representing the details of an update request.

VpcConfigRequest
An object representing the VPC configuration to use for an Amazon EKS cluster.

VpcConfigResponse
An object representing an Amazon EKS cluster VPC configuration response.

Enums

CreateClusterError
Errors returned by CreateCluster

DeleteClusterError
Errors returned by DeleteCluster

DescribeClusterError
Errors returned by DescribeCluster

DescribeUpdateError
Errors returned by DescribeUpdate

ListClustersError
Errors returned by ListClusters

ListUpdatesError
Errors returned by ListUpdates

UpdateClusterConfigError
Errors returned by UpdateClusterConfig

UpdateClusterVersionError
Errors returned by UpdateClusterVersion

Traits

Eks
Trait representing the capabilities of the Amazon EKS API. Amazon EKS clients implement this trait.