Struct rusoto_eks::EksClient

pub struct EksClient { /* fields omitted */ }

A client for the Amazon EKS API.

Methods

impl EksClient

pub fn new(region: Region) -> EksClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> EksClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,