Crate rusoto_eks

Amazon Elastic Container Service for Kubernetes (Amazon EKS) is a managed service that makes it easy for you to run Kubernetes on AWS without needing to stand up or maintain your own Kubernetes control plane. Kubernetes is an open-source system for automating the deployment, scaling, and management of containerized applications.

Amazon EKS runs three Kubernetes control plane instances across three Availability Zones to ensure high availability. Amazon EKS automatically detects and replaces unhealthy control plane instances, and it provides automated version upgrades and patching for them.

Amazon EKS is also integrated with many AWS services to provide scalability and security for your applications, including the following:

Elastic Load Balancing for load distribution

IAM for authentication

Amazon VPC for isolation

Amazon EKS runs up to date versions of the open-source Kubernetes software, so you can use all the existing plugins and tooling from the Kubernetes community. Applications running on Amazon EKS are fully compatible with applications running on any standard Kubernetes environment, whether running in on-premises data centers or public clouds. This means that you can easily migrate any standard Kubernetes application to Amazon EKS without any code modification required.

If you're using the service, you're probably looking for EksClient and Eks.

Structs

Certificate
An object representing the certificate-authority-data for your cluster.

Cluster
An object representing an Amazon EKS cluster.

CreateClusterRequest

CreateClusterResponse

DeleteClusterRequest

DeleteClusterResponse

DescribeClusterRequest

DescribeClusterResponse

EksClient
A client for the Amazon EKS API.

ListClustersRequest

ListClustersResponse

VpcConfigRequest
An object representing an Amazon EKS cluster VPC configuration request.

VpcConfigResponse
An object representing an Amazon EKS cluster VPC configuration response.

Enums

CreateClusterError
Errors returned by CreateCluster

DeleteClusterError
Errors returned by DeleteCluster

DescribeClusterError
Errors returned by DescribeCluster

ListClustersError
Errors returned by ListClusters

Traits

Eks
Trait representing the capabilities of the Amazon EKS API. Amazon EKS clients implement this trait.