Trait rusoto_eks::Eks

pub trait Eks {
    fn create_cluster(
        &self, 
        input: CreateClusterRequest
    ) -> RusotoFuture<CreateClusterResponse, CreateClusterError>;

    fn delete_cluster(
        &self, 
        input: DeleteClusterRequest
    ) -> RusotoFuture<DeleteClusterResponse, DeleteClusterError>;

    fn describe_cluster(
        &self, 
        input: DescribeClusterRequest
    ) -> RusotoFuture<DescribeClusterResponse, DescribeClusterError>;

    fn list_clusters(
        &self, 
        input: ListClustersRequest
    ) -> RusotoFuture<ListClustersResponse, ListClustersError>;
}

Trait representing the capabilities of the Amazon EKS API. Amazon EKS clients implement this trait.

Required Methods

fn create_cluster(
&self,
input: CreateClusterRequest
) -> RusotoFuture<CreateClusterResponse, CreateClusterError>

Creates an Amazon EKS control plane.

The Amazon EKS control plane consists of control plane instances that run the Kubernetes software, like etcd and the API server. The control plane runs in an account managed by AWS, and the Kubernetes API is exposed via the Amazon EKS API server endpoint.

Amazon EKS worker nodes run in your AWS account and connect to your cluster's control plane via the Kubernetes API server endpoint and a certificate file that is created for your cluster.

The cluster control plane is provisioned across multiple Availability Zones and fronted by an Elastic Load Balancing Network Load Balancer. Amazon EKS also provisions elastic network interfaces in your VPC subnets to provide connectivity from the control plane instances to the worker nodes (for example, to support kubectl exec, logs, and proxy data flows).

After you create an Amazon EKS cluster, you must configure your Kubernetes tooling to communicate with the API server and launch worker nodes into your cluster. For more information, see Managing Cluster Authentication and Launching Amazon EKS Worker Nodesin the Amazon EKS User Guide.

fn delete_cluster(
&self,
input: DeleteClusterRequest
) -> RusotoFuture<DeleteClusterResponse, DeleteClusterError>

Deletes the Amazon EKS cluster control plane.

If you have active services in your cluster that are associated with a load balancer, you must delete those services before deleting the cluster so that the load balancers are deleted properly. Otherwise, you can have orphaned resources in your VPC that prevent you from being able to delete the VPC. For more information, see Deleting a Cluster in the Amazon EKS User Guide.

fn describe_cluster(
&self,
input: DescribeClusterRequest
) -> RusotoFuture<DescribeClusterResponse, DescribeClusterError>

Returns descriptive information about an Amazon EKS cluster.

The API server endpoint and certificate authority data returned by this operation are required for kubelet and kubectl to communicate with your Kubernetes API server. For more information, see Create a kubeconfig for Amazon EKS.

The API server endpoint and certificate authority data are not available until the cluster reaches the ACTIVE state.

fn list_clusters(
&self,
input: ListClustersRequest
) -> RusotoFuture<ListClustersResponse, ListClustersError>

Lists the Amazon EKS clusters in your AWS account in the specified region.